pyproject.toml
[tool.poetry]
name = "skipy"
version = "0.1.0"
description = "Skip-tac python modules"
authors = ["syamaguc <programmer2844@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
requests = "^2.31.0"
selenium = "^4.15.2"
boto3 = "^1.29.6"
google-api-python-client = "^2.108.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.11.0"
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
tox = "^4.11.3"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"
mypy = "^1.7.1"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
